+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added Manufacturer tag to metadata.
 * Added mask_xy_corners and mask_xy_ellipse functions.
 * Added functions to fit T1 and T2 relaxation curves.
+* Updated CI checks.
 
 # spant 2.6.0
 * Masked spectra plotted with image are now shown with a value of zero.