+++ NEWS.md
@@ -1,3 +1,6 @@
+# spant 2.8.0
+* Bug fix when using align function with masked data.
+
 # spant 2.7.0
 * Added Manufacturer tag to metadata.
 * Added mask_xy_corners and mask_xy_ellipse functions.